pyproject.toml
[tool.poetry]
name = "oleg-bot"
version = "4.0.0"
description = "Цифровой гигачад — Telegram-бот с ИИ, модерацией и игровыми механиками"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "app" },
]
repository = "https://github.com/your-repo/oleg-bot"
keywords = ["telegram", "bot", "aiogram", "ollama", "ai"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Natural Language :: Russian",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
]

[tool.poetry.dependencies]
python = "~3.12"
aiogram = "^3.13.1"
python-dotenv = "^1.0.1"
SQLAlchemy = "^2.0.36"
aiosqlite = "^0.20.0"
alembic = "^1.14.0"
httpx = "^0.27.2"
APScheduler = "^3.10.4"
pytz = "^2024.2"
python-dateutil = "^2.9.0"
uvloop = { version = "^0.21.0", markers = "sys_platform != 'win32'" }
Pillow = "^10.0.0"
yt-dlp = "^2023.10.13"
chromadb = "^0.6.0"
ollama = "^0.3.3"
pydantic-settings = "^2.12.0"
arq = "^0.26.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.13.2"
mypy = "^1.7.1"

[tool.poetry.scripts]
oleg = "app.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
skip_glob = ["*/migrations/*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
ignore_missing_imports = true
